requirements.txt
apache-airflow[postgres,celery,redis]==2.6.3
requests==2.31.0
pandas==1.5.3
unidecode==1.3.6
matplotlib==3.7.2
seaborn==0.12.2